ch-api-versioning-173105234(versioning): - ensure the routes are versioned [Finishes #173105234]
