+++ routes/routes.go
@@ -1,12 +1,7 @@
 package routes
 
 import (
-	"encoding/json"
-	"github.com/maxwellgithinji/farmsale_backend/controllers/productscontroller"
-	"github.com/maxwellgithinji/farmsale_backend/controllers/userscontroller"
-	"github.com/maxwellgithinji/farmsale_backend/middleware/auth"
 	"github.com/maxwellgithinji/farmsale_backend/middleware/common"
-	"github.com/maxwellgithinji/farmsale_backend/utils"
 	"net/http"
 	_ "net/http/pprof" // For dev only, dont push to production
 
@@ -15,86 +10,21 @@ import (
 
 func RouteHandlers() *mux.Router {
 	r := mux.NewRouter().StrictSlash(true)
-	r.Use(common.CommonMiddleware)
 
-	r.Handle("/favicon.ico", http.NotFoundHandler()).Methods("GET")
+	var api = r.PathPrefix("/api").Subrouter()
 
-	r.HandleFunc("/", index).Methods("GET")
-	r.HandleFunc("/signup", userscontroller.Signup).Methods("POST")
-	r.HandleFunc("/login", userscontroller.Login).Methods("POST")
+	api.NotFoundHandler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.WriteHeader(http.StatusNotFound)
+	})
 
-	//Auth Route
-	s := r.PathPrefix("/auth").Subrouter()
-	s.Use(auth.JwtVerify)
-	s.HandleFunc("/products", productscontroller.Index).Methods("GET")
-	s.HandleFunc("/profile/{id}", userscontroller.EditProfile).Methods("PUT")
+	api.Handle("/favicon.ico", http.NotFoundHandler()).Methods("GET")
+	api.Use(common.CommonMiddleware)
 
-	//Admin Route
-	a := r.PathPrefix("/admin").Subrouter()
-	a.Use(auth.AdminVerify)
-	a.HandleFunc("/", admin).Methods("GET")
-	a.HandleFunc("/profile/delete/{id}", userscontroller.DeleteUser).Methods("DELETE")
-	a.HandleFunc("/profile/activate/{id}", userscontroller.ActivateDeactivateAccount).Methods("PUT")
+	//API V1
+	apiV1(api)
 
-	//Manager Route
-	m := r.PathPrefix("/manager").Subrouter()
-	m.Use(auth.ManagerVerify)
-	m.HandleFunc("/", manager).Methods("GET")
-
-	//Agent Route
-	ag := r.PathPrefix("/agent").Subrouter()
-	ag.Use(auth.AgentVerify)
-	ag.HandleFunc("/", agent).Methods("GET")
-
-	//Current user route
-	cu := r.PathPrefix("/profile").Subrouter()
-	cu.Use(auth.CurrentUserVerify)
-	cu.HandleFunc("/{id}", userscontroller.EditProfile).Methods("PUT")
-	cu.HandleFunc("/deactivate/{id}", userscontroller.DeactivateAccount).Methods("PUT")
+	//API V2 can go here
+	// apiV2(api)
 
 	return r
-}
-
-func index(w http.ResponseWriter, req *http.Request) {
-	if req.Method != http.MethodGet {
-		http.Error(w, http.StatusText(405), 405)
-		return
-	}
-	msg := utils.MessageResponse{
-		Msg: "User welcome to farmsale",
-	}
-	json.NewEncoder(w).Encode(msg)
-}
-
-func admin(w http.ResponseWriter, req *http.Request) {
-	if req.Method != http.MethodGet {
-		http.Error(w, http.StatusText(405), 405)
-		return
-	}
-	msg := utils.MessageResponse{
-		Msg: "Admin welcome to farmsale",
-	}
-	json.NewEncoder(w).Encode(msg)
-}
-
-func agent(w http.ResponseWriter, req *http.Request) {
-	if req.Method != http.MethodGet {
-		http.Error(w, http.StatusText(405), 405)
-		return
-	}
-	msg := utils.MessageResponse{
-		Msg: "Agent welcome to farmsale",
-	}
-	json.NewEncoder(w).Encode(msg)
-}
-
-func manager(w http.ResponseWriter, req *http.Request) {
-	if req.Method != http.MethodGet {
-		http.Error(w, http.StatusText(405), 405)
-		return
-	}
-	msg := utils.MessageResponse{
-		Msg: "Manager welcome to farmsale",
-	}
-	json.NewEncoder(w).Encode(msg)
 }
